修改一些已知的BUG

+++ model/level.go
@@ -22,7 +22,7 @@ type Level struct {
 type Virus struct {
 	Time int `json:"time" comment:"病毒出现时间"`
 	HP int `json:"hp" comment:"病毒HP"`
-	Speed int `json:"speed" comment:"病毒移动速度"`
+	Speed float64 `json:"speed" comment:"病毒移动速度"`
 	X int `json:"x" comment:"病毒出现的X位置"`
 	Y int `json:"y" comment:"病毒出现的Y位置"`
 	Width int `json:"w" comment:"病毒宽度"`
